--- a/rebar.config
+++ b/rebar.config
@@ -58,6 +58,7 @@
 ]}.
 
 {plugins, [
+    % {rebar3_bench, "0.2.1"},
     {erlfmt, "0.8.0"}
 ]}.