--- a/rebar.config
+++ b/rebar.config
@@ -50,7 +50,11 @@
         race_conditions,
         unknown
     ]},
-    {plt_apps, all_deps}
+    {plt_apps, all_deps},
+    {plt_extra_apps, [
+        syntax_tools,
+        compiler
+    ]}
 ]}.
 
 {pre_hooks, [