--- a/src/dmt_domain.erl
+++ b/src/dmt_domain.erl
@@ -356,14 +356,17 @@ references({Tag, Object}, StructInfo = {struct, union, FieldsInfo}, Refs) when i
         {_, _, Type, _, _} ->
             check_reference_type(Object, Type, Refs)
     end;
-%% what if it's a union?
-references(Object, {struct, struct, FieldsInfo}, Refs) when is_list(FieldsInfo) ->
-    lists:foldl(
-        fun({I, {_, _Required, FieldType, _Name, _}}, Acc) ->
-            check_reference_type(element(I, Object), FieldType, Acc)
+references(Object, {struct, struct, FieldsInfo}, Refs) when is_list(FieldsInfo) -> %% what if it's a union?
+    indexfold(
+        fun
+            (I, {_, _Required, FieldType, _Name, _}, Acc) ->
+                check_reference_type(element(I, Object), FieldType, Acc)
         end,
         Refs,
-        mark_fields(FieldsInfo)
+        % NOTE
+        % This `2` gives index of the first significant field in a record tuple.
+        2,
+        FieldsInfo
     );
 references(Object, {struct, _, {?DOMAIN, StructName}}, Refs) ->
     StructInfo = get_struct_info(StructName),
@@ -387,6 +390,11 @@ references(Object, {map, KeyType, ValueType}, Refs) ->
     );
 references(_DomainObject, _Primitive, Refs) ->
     Refs.
+
+indexfold(Fun, Acc, I, [E | Rest]) ->
+    indexfold(Fun, Fun(I, E, Acc), I + 1, Rest);
+indexfold(_Fun, Acc, _I, []) ->
+    Acc.
 
 check_reference_type(undefined, _, Refs) ->
     Refs;
@@ -426,19 +434,20 @@ get_field_info(Field, {struct, _StructType, FieldsInfo}) ->
     lists:keyfind(Field, 4, FieldsInfo).
 
 get_field_index(Field, {struct, _StructType, FieldsInfo}) ->
-    get_field_index(Field, mark_fields(FieldsInfo));
-get_field_index(_Field, []) ->
-    false;
-get_field_index(Field, [F | Rest]) ->
-    case F of
-        {_, {_, _, _, Field, _}} = Index ->
-            Index;
-        _ ->
-            get_field_index(Field, Rest)
-    end.
+    % NOTE
+    % This `2` gives index of the first significant field in a record tuple.
+    get_field_index(Field, 2, FieldsInfo).
 
-mark_fields(FieldsInfo) ->
-    lists:zip(lists:seq(2, 1 + length(FieldsInfo)), FieldsInfo).
+get_field_index(_Field, _, []) ->
+    false;
+
+get_field_index(Field, I, [F | Rest]) ->
+    case F of
+        {_, _, _, Field, _} = Info ->
+            {I, Info};
+        _ ->
+            get_field_index(Field, I + 1, Rest)
+    end.
 
 is_reference_type(Type) ->
     {struct, union, StructInfo} = get_struct_info('Reference'),