--- a/src/dmt_domain_pt.erl
+++ b/src/dmt_domain_pt.erl
@@ -1,0 +1,65 @@
+-module(dmt_domain_pt).
+
+-export([parse_transform/2]).
+
+-spec parse_transform(Forms, [compile:option()]) -> Forms when
+    Forms :: [erl_parse:abstract_form() | erl_parse:form_info()].
+parse_transform(Forms, _Options) ->
+    [
+        erl_syntax:revert(FormNext) ||
+            Form <- Forms,
+                FormNext <- [erl_syntax_lib:map(fun transform/1, Form)],
+                    FormNext /= delete
+    ].
+
+transform(Form) ->
+    case erl_syntax:type(Form) of
+        function ->
+            Name = erl_syntax:concrete(erl_syntax:function_name(Form)),
+            Arity = erl_syntax:function_arity(Form),
+            transform_function(Name, Arity, Form);
+        _ ->
+            Form
+    end.
+
+transform_function(Name = is_reference_type, 1, FormWas) ->
+    % NOTE
+    % Replacing `dmt_domain:is_reference_type/1` with a code which does something similar to:
+    % ```
+    % is_reference_type({struct, struct, {dmsl_domain_thrift, 'CategoryRef'}}) -> {true, 'category'};
+    % is_reference_type({struct, struct, {dmsl_domain_thrift, 'CurrencyRef'}}) -> {true, 'currency'};
+    % ...
+    % is_reference_type(_) -> false.
+    % ```
+    {struct, union, StructInfo} = get_struct_info('Reference'),
+    Clauses = [
+        erl_syntax:clause(
+            [erl_syntax:abstract(Type)],
+            none,
+            [erl_syntax:abstract({true, Tag})]
+        ) || {_N, _Req, Type, Tag, _Default} <- StructInfo
+    ] ++ [
+        erl_syntax:clause(
+            [erl_syntax:underscore()],
+            none,
+            [erl_syntax:abstract(false)]
+        )
+    ],
+    Form = erl_syntax_lib:map(
+        fun (F) -> erl_syntax:copy_attrs(FormWas, F) end,
+        erl_syntax:function(
+            erl_syntax:abstract(Name),
+            Clauses
+        )
+    ),
+    Form;
+transform_function(_Name = is_reference_type, 2, _FormWas) ->
+    % NOTE
+    % We need to make `is_reference_type/2` disappear, otherwise it will trigger _unused function_
+    % warning.
+    delete;
+transform_function(_, _, Form) ->
+    Form.
+
+get_struct_info(StructName) ->
+    dmsl_domain_thrift:struct_info(StructName).