## Workload

Snapshot version: 23490
Snapshot size: 5664
Number of commits: 3
Number of operations: 102
Number of insertions: 0
Number of updates: 0
Number of removals: 102

## Results

Numbers after each iteration. I had to remeasure **2** under laptop battery power setting, that's why numbers afterwards look a bit higher.

### 1. No mark_fields (10fde55)

Compared to **baseline** on master.

```
===> Stats for wall_time
Min:                721.62ms (-   37.81ms /  5.0%)
25 percentile:      727.49ms (-   36.58ms /  4.8%)
Median:             731.06ms (-   38.02ms /  4.9%)
75 percentile:      732.70ms (-   42.69ms /  5.5%)
Max:                769.08ms (-   48.19ms /  5.9%)
Outliers:         Lo: 0; Hi: 1; Sum: 1
Outlier variance: 4.75000e-2 (slight)
> Bootstrapped
Mean:               732.29ms (-   40.91ms /  5.3%)
Std deviation:     9800.67μs
> Relative
Difference at 95.0 confidence
   -40.91ms ±  7686.70μs
    -5.29%  ±  0.99%
 (Student's t, pooled s = 1.20096e+7
```

### 2. Straightforward maps:fold (ef3a987)

Compared to **1** profile.

```
===> Stats for wall_time
Min:                710.96ms (-   11.34ms /  1.6%)
25 percentile:      714.78ms (-   14.62ms /  2.0%)
Median:             731.30ms (- 7843.70μs /  1.1%)
75 percentile:      740.59ms (- 2033.11μs /  0.3%)
Max:                749.82ms (-   38.80ms /  4.9%)
Outliers:         Lo: 0; Hi: 0; Sum: 0
Outlier variance: 4.75000e-2 (slight)
> Bootstrapped
Mean:               729.58ms (-   12.54ms /  1.7%)
Std deviation:       12.13ms
> Relative
Difference at 95.0 confidence
   -12.54ms ±  9383.82μs
    -1.69%  ±  1.26%
 (Student's t, pooled s = 1.46612e+7
```

### 3. Less useless `undefined` matching (1b255a2)

Compared to **2**, measured under _battery_ setting, hence the numbers.

Even though no statistically significant improvement was observed, results were consistent through several runs.

```
===> Stats for wall_time
Min:                710.07ms (- 3943.33μs /  0.6%)
25 percentile:      719.05ms (- 3650.90μs /  0.5%)
Median:             727.46ms (- 1068.10μs /  0.1%)
75 percentile:      740.76ms (- 1397.55μs /  0.2%)
Max:                747.87ms (-   80.81μs /  0.0%)
Outliers:         Lo: 0; Hi: 0; Sum: 0
Outlier variance: 4.75000e-2 (slight)
> Bootstrapped
Mean:               730.28ms (- 1000.93μs /  0.1%)
Std deviation:       11.67ms
> Relative
No difference found at 95.0% confidence level
```

### 4. With precomputed reftype mapping (5176848)

Compared to **2**, measured under _battery_ setting.

```
===> Stats for wall_time
Min:                440.27ms (-  273.74ms / 38.3%)
25 percentile:      446.27ms (-  276.43ms / 38.3%)
Median:             454.35ms (-  274.18ms / 37.6%)
75 percentile:      464.81ms (-  277.34ms / 37.4%)
Max:                471.85ms (-  276.10ms / 36.9%)
Outliers:         Lo: 0; Hi: 0; Sum: 0
Outlier variance: 4.75000e-2 (slight)
> Bootstrapped
Mean:               455.99ms (-  275.29ms / 37.6%)
Std deviation:       11.32ms
> Relative
Difference at 95.0 confidence
  -275.29ms ±  7372.71μs
   -37.65%  ±  1.01%
 (Student's t, pooled s = 1.15190e+7
```

### 5. With parse transform hack (aa957fa)

Compared to **2**, measured under _battery_ setting.

```
===> Stats for wall_time
Min:                120.71ms (-  593.30ms / 83.1%)
25 percentile:      121.59ms (-  601.11ms / 83.2%)
Median:             123.04ms (-  605.49ms / 83.1%)
75 percentile:      123.62ms (-  618.54ms / 83.3%)
Max:                127.29ms (-  620.66ms / 83.0%)
Outliers:         Lo: 0; Hi: 1; Sum: 1
Outlier variance: 4.75000e-2 (slight)
> Bootstrapped
Mean:               122.97ms (-  608.31ms / 83.2%)
Std deviation:     1576.56μs
> Relative
Difference at 95.0 confidence
  -608.31ms ±  5351.30μs
   -83.18%  ±  0.73%
 (Student's t, pooled s = 8.36082e+6
```